requirements.txt
numpy==2.3.2
openai==1.101.0
openai_whisper==20250625
python-dotenv==1.1.1
scikit_learn==1.7.1
sentence_transformers==5.1.0
streamlit==1.48.1
tqdm==4.67.1
whisper==1.1.10
